requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
lightning-utilities==0.15.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.3.3
omegaconf==2.3.0
packaging==25.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pytorch-lightning==2.5.5
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.37.1
setuptools==80.9.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.4.0
torch-geometric==2.6.1
torch_cluster==1.6.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torchmetrics==1.8.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.21.4
yarl==1.20.1
